+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Use item id instead of barcode for links on request details screen. Fixes UIREQ-484.
 * Refactor to `miragejs` from `@bigtest/mirage`.
 * Add loading indicator when service point is switched. Fixes UIREQ-508.
+* Use patron group name ('group') instead of description in displays. Fixes UIREQ-499.
 
 ## [3.0.1](https://github.com/folio-org/ui-requests/tree/v3.0.1) (2020-06-19)
 [Full Changelog](https://github.com/folio-org/ui-requests/compare/v3.0.0...v3.0.1)